feature/api - implemented DoctorPolicy class and rules for doctor's method

+++ app/Http/Controllers/DoctorController.php
@@ -45,6 +45,8 @@ class DoctorController extends Controller
 
     public function destroy(Doctor $doctor, DoctorService $service): void
     {
+        $this->authorize('delete', $doctor);
+
         $service->destroy($doctor);
     }
 }
